backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
pydantic==2.9.0
aiosqlite==0.20.0
Pillow==10.4.0
numpy
scipy
opencv-python-headless
librosa==0.10.2
soundfile==0.12.1
torch
torchvision
tensorflow==2.20.0
scikit-learn
httpx
python-dotenv
spacy>=3.8.0
